+++ administrator/components/com_virtuemart/models/manufacturercategories.php
@@ -13,13 +13,13 @@
 * to the GNU General Public License, and as distributed it includes or
 * is derivative of works licensed under the GNU General Public License or
 * other free or open source software licenses.
-* @version $Id: manufacturercategories.php 6350 2012-08-14 17:18:08Z Milbo $
+* @version $Id: manufacturercategories.php 8953 2015-08-19 10:30:52Z Milbo $
 */
 
 // Check to ensure this file is included in Joomla!
 defined('_JEXEC') or die('Restricted access');
 
-if(!class_exists('VmModel'))require(JPATH_VM_ADMINISTRATOR.DS.'helpers'.DS.'vmmodel.php');
+if(!class_exists('VmModel'))require(VMPATH_ADMIN.DS.'helpers'.DS.'vmmodel.php');
 
 /**
  * Model class for manufacturer category
@@ -40,55 +40,42 @@ class VirtuemartModelManufacturercategories extends VmModel {
 		parent::__construct('virtuemart_manufacturercategories_id');
 		$this->setMainTable('manufacturercategories');
 		$this->addvalidOrderingFieldName(array('mf_category_name'));
-		$config=JFactory::getConfig();
 	}
 
-    /**
-     * Retrieve the detail record for the current $id if the data has not already been loaded.
-     *
-     */
-	// function getManufacturerCategory(){
-
-		//// $db = JFactory::getDBO();
-
-  		// if (empty($this->_data)) {
-   			// $this->_data = $this->getTable('manufacturercategories');
-   			// $this->_data->load((int)$this->_id);
-  		// }
-
-		//// print_r( $this->_db->_sql );
-  		// if (!$this->_data) {
-   			// $this->_data = new stdClass();
-   			// $this->_id = 0;
-   			// $this->_data = null;
-  		// }
-
-  		// return $this->_data;
-	// }
 	/**
 	 * Delete all record ids selected
      *
      * @return boolean True is the remove was successful, false otherwise.
      */
-	function remove($categoryIds)
-	{
+	function remove($categoryIds) {
+		if(!vmAccess::manager('manufacturercategories')){
+			vmWarn('Insufficient permissions to delete manufacturer category');
+			return false;
+		}
+
     	$table = $this->getTable('manufacturercategories');
 
     	foreach($categoryIds as $categoryId) {
        		if($table->checkManufacturer($categoryId)) {
 	    		if (!$table->delete($categoryId)) {
-	            		vmError($table->getError());
 	            		return false;
 	       		}
        		}
        		else {
-				vmError(get_class( $this ).'::remove '.$categoryId.' '.$table->getError());
+				vmError(get_class( $this ).'::remove '.$categoryId.' failed');
        			return false;
        		}
     	}
     	return true;
 	}
 
+	function store(&$data){
+		if(!vmAccess::manager('manufacturercategories')){
+			vmWarn('Insufficient permissions to store manufacturer category');
+			return false;
+		}
+		return parent::store($data);
+	}
 
 	/**
 	 * Retireve a list of countries from the database.
@@ -100,7 +87,7 @@ class VirtuemartModelManufacturercategories extends VmModel {
 	function getManufacturerCategories($onlyPublished=false, $noLimit=false)
 	{
 		$this->_noLimit = $noLimit;
-		$select = '* FROM `#__virtuemart_manufacturercategories_'.VMLANG.'` as l';
+		$select = '* FROM `#__virtuemart_manufacturercategories_'.VmConfig::$vmlang.'` as l';
 		$joinedTables = ' JOIN `#__virtuemart_manufacturercategories` as mc using (`virtuemart_manufacturercategories_id`)';
 		$where = array();
 		if ($onlyPublished) {
@@ -111,7 +98,7 @@ class VirtuemartModelManufacturercategories extends VmModel {
 
 		$whereString = '';
 		if (count($where) > 0) $whereString = ' WHERE '.implode(' AND ', $where) ;
-		if ( JRequest::getCmd('view') == 'manufacturercategories') {
+		if ( vRequest::getCmd('view') == 'manufacturercategories') {
 			$ordering = $this->_getOrdering();
 		} else {
 			$ordering = ' order by mf_category_name DESC';
@@ -128,13 +115,11 @@ class VirtuemartModelManufacturercategories extends VmModel {
 	function getCategoryFilter(){
 		$db = JFactory::getDBO();
 		$query = 'SELECT `virtuemart_manufacturercategories_id` as `value`, `mf_category_name` as text'
-				.' FROM #__virtuemart_manufacturercategories_'.VMLANG.'`';
+				.' FROM `#__virtuemart_manufacturercategories_'.VmConfig::$vmlang.'`';
 		$db->setQuery($query);
 
-		$categoryFilter[] = JHTML::_('select.option',  '0', '- '. JText::_('COM_VIRTUEMART_SELECT_MANUFACTURER_CATEGORY') .' -' );
-
+		$categoryFilter[] = JHtml::_('select.option',  '0', '- '. vmText::_('COM_VIRTUEMART_SELECT_MANUFACTURER_CATEGORY') .' -' );
 		$categoryFilter = array_merge($categoryFilter, (array)$db->loadObjectList());
-
 
 		return $categoryFilter;
